+++ chaoslib/types.py
@@ -4,7 +4,7 @@ from typing import Any, Dict, Generator, List, Tuple, Union
 __all__ = ["MicroservicesStatus", "Probe", "Action", "Experiment", "Layer",
            "TargetLayers", "Activity", "Journal", "Run", "Secrets", "Step",
            "Configuration", "Discovery", "DiscoveredActivities",
-           "DiscoveredSystemInfo", "Settings", "EventPayload"]
+           "DiscoveredSystemInfo", "Settings", "EventPayload", "Tolerance"]
 
 
 Action = Dict[str, Any]
@@ -30,3 +30,5 @@ DiscoveredSystemInfo = Dict[str, Any]
 
 Settings = Dict[str, Any]
 EventPayload = Dict[str, Any]
+
+Tolerance = Union[int, str, bool, list, Dict[str, Any]]